requirements.txt
accelerate==0.14.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bpemb==0.3.4
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
conllu==4.5.2
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.28
datasets==2.6.1
deepl==1.11.0
Deprecated==1.2.13
dill==0.3.5.1
evaluate==0.3.0
filelock==3.8.0
flair==0.11.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.2
gdown==4.4.0
gensim==4.2.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.10.1
hyperframe==5.2.0
hyperopt==0.2.7
idna==2.10
importlib-metadata==3.10.1
Janome==0.4.2
joblib==1.2.0
kiwisolver==1.4.4
konoha==4.6.5
langdetect==1.0.9
lxml==4.9.1
matplotlib==3.6.2
more-itertools==9.0.0
mpld3==0.3
multidict==6.0.2
multiprocess==0.70.13
networkx==2.8.8
nltk==3.7
numpy==1.23.4
overrides==3.1.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pptree==3.1
psutil==5.9.4
py4j==0.10.9.7
pyarrow==10.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
responses==0.18.0
rfc3986==1.5.0
scikit-learn==1.1.3
scipy==1.9.3
segtok==1.5.11
sentencepiece==0.1.95
six==1.16.0
sklearn==0.0.post1
smart-open==6.2.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlitedict==2.0.0
tabulate==0.9.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Wikipedia-API==0.5.4
wrapt==1.14.1
xxhash==3.1.0
yarl==1.8.1
zipp==3.11.0
